+++ samples/client/petstore/python/swagger_client/apis/user_api.py
@@ -2,7 +2,7 @@
 
 """
 UserApi.py
-Copyright 2015 SmartBear Software
+Copyright 2016 SmartBear Software
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -79,6 +79,7 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/user'.replace('{format}', 'json')
         method = 'POST'
 
@@ -88,7 +89,7 @@ class UserApi(object):
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -154,6 +155,7 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/user/createWithArray'.replace('{format}', 'json')
         method = 'POST'
 
@@ -163,7 +165,7 @@ class UserApi(object):
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -229,6 +231,7 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/user/createWithList'.replace('{format}', 'json')
         method = 'POST'
 
@@ -238,7 +241,7 @@ class UserApi(object):
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -305,6 +308,7 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/user/login'.replace('{format}', 'json')
         method = 'GET'
 
@@ -318,7 +322,7 @@ class UserApi(object):
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -381,6 +385,7 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/user/logout'.replace('{format}', 'json')
         method = 'GET'
 
@@ -390,7 +395,7 @@ class UserApi(object):
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -440,9 +445,6 @@ class UserApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'username' is set
-        if username is None:
-            raise ValueError("Missing the required parameter `username` when calling `get_user_by_name`")
 
         all_params = ['username']
         all_params.append('callback')
@@ -457,6 +459,10 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+        # verify the required parameter 'username' is set
+        if ('username' not in params) or (params['username'] is None):
+            raise ValueError("Missing the required parameter `username` when calling `get_user_by_name`")
+
         resource_path = '/user/{username}'.replace('{format}', 'json')
         method = 'GET'
 
@@ -468,7 +474,7 @@ class UserApi(object):
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -519,9 +525,6 @@ class UserApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'username' is set
-        if username is None:
-            raise ValueError("Missing the required parameter `username` when calling `update_user`")
 
         all_params = ['username', 'body']
         all_params.append('callback')
@@ -536,6 +539,10 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+        # verify the required parameter 'username' is set
+        if ('username' not in params) or (params['username'] is None):
+            raise ValueError("Missing the required parameter `username` when calling `update_user`")
+
         resource_path = '/user/{username}'.replace('{format}', 'json')
         method = 'PUT'
 
@@ -547,7 +554,7 @@ class UserApi(object):
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
@@ -599,9 +606,6 @@ class UserApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'username' is set
-        if username is None:
-            raise ValueError("Missing the required parameter `username` when calling `delete_user`")
 
         all_params = ['username']
         all_params.append('callback')
@@ -616,6 +620,10 @@ class UserApi(object):
             params[key] = val
         del params['kwargs']
 
+        # verify the required parameter 'username' is set
+        if ('username' not in params) or (params['username'] is None):
+            raise ValueError("Missing the required parameter `username` when calling `delete_user`")
+
         resource_path = '/user/{username}'.replace('{format}', 'json')
         method = 'DELETE'
 
@@ -627,7 +635,7 @@ class UserApi(object):
 
         header_params = {}
 
-        form_params = {}
+        form_params = []
         files = {}
 
         body_params = None
